Move theme to shared directory
