+++ src/components/rich-text-editor/style.js
@@ -3,7 +3,7 @@ import React from 'react';
 import styled, { css } from 'styled-components';
 import Link from 'src/components/link';
 import { Transition, zIndex } from '../globals';
-import { theme } from 'src/components/theme';
+import { theme } from 'shared/theme';
 
 export const customStyleMap = {
   CODE: {
